featL fft cleanup
